Revert "Modularize wasm account (#789)"
This reverts commit 84114bbd3e548ee3108e6663ef0f61d12ebf470f.
