+++ packages/account-wasm/src/types/mod.rs
@@ -11,7 +11,6 @@ pub(crate) mod invocation;
 pub(crate) mod outside_execution;
 pub(crate) mod policy;
 pub(crate) mod session;
-pub(crate) mod signer;
 
 #[allow(non_snake_case)]
 #[derive(Tsify, Serialize, Deserialize, Debug, Clone)]
